Initial fix. Needs testing
Added a set of "affected" dst registers for each gadget, which causes createDependenceChain to add an extra gadget at the end to
fix an earlier pop being overwritten. This isn't optimal but for now it's better than creating a non-functional chain.

Next issue: it seems that the triple-pop gadget being chosen (which was chosen before as well) is being fed the padding/actual value in the
wrong order.